+++ python_modules/dagster/dagster/_core/utils.py
@@ -3,7 +3,6 @@ import random
 import re
 import string
 import uuid
-import warnings
 from collections import OrderedDict, deque
 from collections.abc import Iterable, Iterator, Mapping, Sequence
 from concurrent.futures import Future, ThreadPoolExecutor, TimeoutError
@@ -24,7 +23,6 @@ from weakref import WeakSet
 import toposort as toposort_
 
 import dagster._check as check
-from dagster._utils import library_version_from_core_version, parse_package_version
 
 BACKFILL_TAG_LENGTH = 8
 
@@ -100,29 +98,6 @@ def str_format_list(items: Iterable[object]) -> str:
 
 def str_format_set(items: Iterable[object]) -> str:
     return "[{items}]".format(items=", ".join([f"'{item}'" for item in items]))
-
-
-def check_dagster_package_version(library_name: str, library_version: str) -> None:
-    # This import must be internal in order for this function to be testable
-    from dagster.version import __version__
-
-    parsed_lib_version = parse_package_version(library_version)
-    if parsed_lib_version.release[0] >= 1:
-        if library_version != __version__:
-            message = (
-                f"Found version mismatch between `dagster` ({__version__})"
-                f"and `{library_name}` ({library_version})"
-            )
-            warnings.warn(message)
-    else:
-        target_version = library_version_from_core_version(__version__)
-        if library_version != target_version:
-            message = (
-                f"Found version mismatch between `dagster` ({__version__}) "
-                f"expected library version ({target_version}) "
-                f"and `{library_name}` ({library_version})."
-            )
-            warnings.warn(message)
 
 
 def get_env_var_name(env_var_str: str):
