+++ python_modules/dagster/dagster_tests/core_tests/test_utils.py
@@ -1,18 +1,12 @@
 import time
-import warnings
 from concurrent.futures import as_completed
 from contextvars import ContextVar
 from typing import NamedTuple
 
-import dagster.version
 import pytest
 from dagster._core.test_utils import environ
-from dagster._core.utils import (
-    InheritContextThreadPoolExecutor,
-    check_dagster_package_version,
-    parse_env_var,
-)
-from dagster._utils import hash_collection, library_version_from_core_version
+from dagster._core.utils import InheritContextThreadPoolExecutor, parse_env_var
+from dagster._utils import hash_collection
 
 
 def test_parse_env_var_no_equals():
@@ -36,40 +30,6 @@ def test_parse_env_var_containing_equals():
     env_var = "FOO_ENV_VAR=HERE_COMES_THE_EQUALS=THERE_IT_WENT"
 
     assert parse_env_var(env_var) == ("FOO_ENV_VAR", "HERE_COMES_THE_EQUALS=THERE_IT_WENT")
-
-
-def test_check_dagster_package_version(monkeypatch):
-    monkeypatch.setattr(dagster.version, "__version__", "1.1.0")
-
-    # Ensure no warning emitted
-    with warnings.catch_warnings():
-        warnings.simplefilter("error")
-        check_dagster_package_version("foo", "1.1.0")
-
-    # Lib version matching 1.1.0-- see dagster._utils.library_version_from_core_version
-    with warnings.catch_warnings():
-        warnings.simplefilter("error")
-        check_dagster_package_version("foo", "0.17.0")
-
-    with pytest.warns(Warning):  # minor version
-        check_dagster_package_version("foo", "1.2.0")
-
-    with pytest.warns(Warning):  # patch version
-        check_dagster_package_version("foo", "1.1.1")
-
-    with pytest.warns(Warning):  # minor version
-        check_dagster_package_version("foo", "0.18.0")
-
-    with pytest.warns(Warning):  # patch version
-        check_dagster_package_version("foo", "0.17.1")
-
-
-def test_library_version_from_core_version():
-    assert library_version_from_core_version("1.1.16") == "0.17.16"
-    assert library_version_from_core_version("0.17.16") == "0.17.16"
-    assert library_version_from_core_version("1.1.16pre0") == "0.17.16rc0"
-    assert library_version_from_core_version("1.1.16rc0") == "0.17.16rc0"
-    assert library_version_from_core_version("1.1.16post0") == "0.17.16post0"
 
 
 def test_hash_collection():
