+++ python_modules/libraries/dagster-dg/dagster_dg/__init__.py
@@ -1,4 +1,5 @@
 from dagster_shared.libraries import DagsterLibraryRegistry
-from dagster_shared.version import __version__ as __version__
+
+from dagster_dg.version import __version__ as __version__
 
 DagsterLibraryRegistry.register("dagster-dg", __version__)
