+++ python_modules/libraries/dagster-shared/tox.ini
@@ -20,4 +20,4 @@ allowlist_externals =
   uv
 commands =
   !windows: /bin/bash -c '! pip list --exclude-editable | grep -e dagster -e dagit'
-  pytest -vv ./dagster_shared_tests
+  pytest -vv ./dagster_shared_tests {posargs}
